Make psk valid in case anyone tries to use the config
The previous psk, "secret", actually was an invalid password and the
wpa_supplicant would error out. Not that that was actually bad since
wifi wasn't going to connect anyway, but it actually would crash
nerves_interim_wifi. The crash has been fixed, but it seems better to
have a validating password anyway.