Collect firmware size statistics
Over the years, we've been adding functionality to the base Nerves
systems to make them easier to use. The consequence is that this has
bloated the images significantly and it's not obvious what to do to
shrink them down again. Additionally, some of the bloat is unintentional
and it would be great to catch that sooner rather than later.

The nerves_examples repository seems like the easiest place to capture
stats on a variety of small projects. This adds firmware size collection
to the CircleCI builds and saves the result to a csv file in the stats
branch.

Processing of the csv file is currently done offline.