Add lock files to support reproducible stats
Without lock files, the projects update themselves on each clean
checkout. This is not desirable at all if we want to reproduce firmware
size stats (or any other stats).

We used to think this was ok since it saved quite a bit of
time maintaining the examples. Back then, we also had to maintain a
lockfile per target and that's been fixed. If updating the lock files is
too laborious, we should be able to automate this by doing a periodic
`mix deps.update --all` and committing if there are any differences.