Update hello_phoenix to work in host mode
The `hello_phoenix` example does not currently work in host mode because
the `firmware`'s local path dependency on `ui` only uses it for targets.
Moreover, the config in the `firmware` project does not work in host
mode against the `ui` dependency.

This commit fixes that by:
  - allowing the `ui` dep to be compiled in host mode as well as for all targets
  - correctly configuring the `ui` app in host mode

It also adds two simple delegate functions, with automated tests, to the
`firmware` project to verify that the `ui` poncho dependency is compiled
and accessible in host mode.

Closes #257.