Prevent unsupported `file_system` dep in firmware
Fixes #386

This is a host-only tool used with `phoenix_live_reload` which can
cause problems if attempting to run on target. In most cases, this
dep doesn't pass the checks during firmware compilation since it
includes host binaries incompatible with the specified target.

This changes that to only compile `:phoenix_live_reload` (and
subsequently `:file_system`) when the target is `:host`